Move Adornments and ContractAdornments projects to source directory
This commit removes the separated Visual Studio 2010 and 2012 projects,
leaving a single project file in the same directory as the source code.
This structure eliminates the need to use linked files in all the projects,
vastly simplifying long-term maintenance.
