+++ CodeContracts.sln
@@ -295,13 +295,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeContractsForRoslyn", "M
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CCDocGenAttributes", "microsoft.research\ManagedContract.Setup\CCDocGenAttributes\CCDocGenAttributes.csproj", "{449B96DF-2C90-4495-B081-D97D4F02620E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adornments", "Microsoft.Research\VSAdornments\VS2010\Adornments.csproj", "{784AA6A9-013E-4D8E-82D6-934571CF1357}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adornments", "Microsoft.Research\VSAdornments\Adornments.csproj", "{784AA6A9-013E-4D8E-82D6-934571CF1357}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adornments11", "Microsoft.Research\VSAdornments\VS2012\Adornments11.csproj", "{235196BB-603A-470A-AF8E-EF6C8D01C4DC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ContractAdornments", "Microsoft.Research\ContractAdornments\VS2010\ContractAdornments.csproj", "{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ContractAdornments11", "Microsoft.Research\ContractAdornments\VS2012\ContractAdornments11.csproj", "{5982130C-B0DD-4E10-82F7-A23F40A16315}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ContractAdornments", "Microsoft.Research\ContractAdornments\Extension\ContractAdornments.csproj", "{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BLResults", "Microsoft.Research\CCTools\BLResults\BLResults.csproj", "{6603246B-9D29-485C-BC38-E8E03F737BC4}"
 EndProject
@@ -2117,27 +2113,6 @@ Global
 		{784AA6A9-013E-4D8E-82D6-934571CF1357}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{784AA6A9-013E-4D8E-82D6-934571CF1357}.Release|Win32.ActiveCfg = Release|Any CPU
 		{784AA6A9-013E-4D8E-82D6-934571CF1357}.Release|x86.ActiveCfg = Release|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Devlab9|Any CPU.ActiveCfg = Devlab9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Devlab9|Mixed Platforms.ActiveCfg = Devlab9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Devlab9|Mixed Platforms.Build.0 = Devlab9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Devlab9|Win32.ActiveCfg = Devlab9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Devlab9|x86.ActiveCfg = Devlab9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Internal9|Any CPU.ActiveCfg = Internal9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Internal9|Mixed Platforms.ActiveCfg = Internal9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Internal9|Mixed Platforms.Build.0 = Internal9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Internal9|Win32.ActiveCfg = Internal9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Internal9|x86.ActiveCfg = Internal9|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Release|Win32.ActiveCfg = Release|Any CPU
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC}.Release|x86.ActiveCfg = Release|Any CPU
 		{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}.Debug|Win32.ActiveCfg = Debug|Any CPU
@@ -2157,26 +2132,6 @@ Global
 		{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}.Release|Win32.ActiveCfg = Release|Any CPU
 		{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C}.Release|x86.ActiveCfg = Release|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Devlab9|Any CPU.ActiveCfg = Devlab9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Devlab9|Mixed Platforms.ActiveCfg = Devlab9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Devlab9|Mixed Platforms.Build.0 = Devlab9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Devlab9|Win32.ActiveCfg = Devlab9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Devlab9|x86.ActiveCfg = Devlab9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Internal9|Any CPU.ActiveCfg = Internal9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Internal9|Mixed Platforms.ActiveCfg = Internal9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Internal9|Mixed Platforms.Build.0 = Internal9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Internal9|Win32.ActiveCfg = Internal9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Internal9|x86.ActiveCfg = Internal9|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Release|Win32.ActiveCfg = Release|Any CPU
-		{5982130C-B0DD-4E10-82F7-A23F40A16315}.Release|x86.ActiveCfg = Release|Any CPU
 		{6603246B-9D29-485C-BC38-E8E03F737BC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6603246B-9D29-485C-BC38-E8E03F737BC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6603246B-9D29-485C-BC38-E8E03F737BC4}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -2424,9 +2379,7 @@ Global
 		{F9E1200C-C624-498A-BCB5-7E55CA1335F4} = {6468A140-0BAF-4813-A1EF-094DD7854C64}
 		{449B96DF-2C90-4495-B081-D97D4F02620E} = {7CD8B829-346E-4FA0-ADDD-87152DDC5DDE}
 		{784AA6A9-013E-4D8E-82D6-934571CF1357} = {7A308E8B-00AB-4ACF-8563-BF457A379EF1}
-		{235196BB-603A-470A-AF8E-EF6C8D01C4DC} = {7A308E8B-00AB-4ACF-8563-BF457A379EF1}
 		{D3F9BE8B-BA9F-4919-87CE-EE1878EDA40C} = {7A308E8B-00AB-4ACF-8563-BF457A379EF1}
-		{5982130C-B0DD-4E10-82F7-A23F40A16315} = {7A308E8B-00AB-4ACF-8563-BF457A379EF1}
 		{6603246B-9D29-485C-BC38-E8E03F737BC4} = {8C9BFE1D-90BD-4F88-B58E-89FF1DA325B8}
 		{CC40F8DC-AA1F-4816-8E2D-18AC2B02E847} = {21B40F99-9AB5-4E2D-AA6E-B9687F5C865A}
 		{407EF6D9-A736-4FEE-A522-789C64E74509} = {8C9BFE1D-90BD-4F88-B58E-89FF1DA325B8}
