+++ Microsoft.Research/ContractAdornments/Extension/AdornmentSource.cs
