+++ Microsoft.Research/ContractAdornments/Extension/Contracts/ContractsProvider.cs
