+++ Microsoft.Research/ContractAdornments/Extension/Delegates.cs
