+++ Microsoft.Research/ContractAdornments/Extension/Extensions.cs
