+++ Microsoft.Research/ContractAdornments/Extension/Inheritance/InheritanceTracker.cs
