+++ Microsoft.Research/ContractAdornments/Extension/Inheritance/MethodCollector.cs
