+++ Microsoft.Research/ContractAdornments/Extension/Inheritance/PropertyCollector.cs
