+++ Microsoft.Research/ContractAdornments/Extension/Intellisense/IntellisenseContractsHelper.cs
