+++ Microsoft.Research/ContractAdornments/Extension/Intellisense/QuickInfo.cs
