+++ Microsoft.Research/ContractAdornments/Extension/Intellisense/QuickInfoProvider.cs
