+++ Microsoft.Research/ContractAdornments/Extension/Intellisense/SignatureHelp.cs
