+++ Microsoft.Research/ContractAdornments/Extension/Intellisense/SignatureHelpProvider.cs
