+++ Microsoft.Research/ContractAdornments/Extension/LineTransformer.cs
