+++ Microsoft.Research/ContractAdornments/Extension/Metadata/MetadataTracker.cs
