+++ Microsoft.Research/ContractAdornments/Extension/Properties/AssemblyInfo.cs
