+++ Microsoft.Research/ContractAdornments/Extension/TextView/TextViewProvider.cs
