+++ Microsoft.Research/ContractAdornments/Extension/TextView/TextViewTracker.cs
