+++ Microsoft.Research/ContractAdornments/Extension/VSServiceProvider/GlobalSuppressions.cs
