+++ Microsoft.Research/ContractAdornments/Extension/VSServiceProvider/Guids.cs
