+++ Microsoft.Research/ContractAdornments/Extension/VSServiceProvider/NonlockingHost.cs
