+++ Microsoft.Research/ContractAdornments/Extension/VSServiceProvider/OptionsPage/OptionsPage.cs
