+++ Microsoft.Research/ContractAdornments/Extension/VSServiceProvider/OptionsPage/OptionsResources.resx
