+++ Microsoft.Research/ContractAdornments/Extension/VSServiceProvider/ProjectTracker.cs
