+++ Microsoft.Research/ContractAdornments/Extension/VSServiceProvider/VSServiceProvider.cs
