+++ Microsoft.Research/ContractAdornments/Extension/source.extension.vsixmanifest
@@ -1,24 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Vsix xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" Version="1.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2010">
   <Identifier Id="ContractAdornments.MicrosoftResearch.1fb66ab7-53ec-479b-bd35-e9dd5be0426b">
-    <Name>Code Contracts Editor Extensions VS2010</Name>
+    <Name>Code Contracts Editor Extensions</Name>
     <Author>Microsoft Research</Author>
     <Version>1.8.10107.10</Version>
-    <Description xml:space="preserve">Displays Code Contracts (when editing C#) in code, Intellisense, and in metadata files.
+    <Description xml:space="preserve">Displays Code Contracts (when editing C#) in code, IntelliSense, and in metadata files.
 
 Inheritance:
-Displays contract information on methods with inherited contract information, or on interface/abstract methods
+Displays contracts on methods with inherited contracts, or on interface/abstract methods
 
 Metadata:
-Displays contract information in metadata files.
+Displays contracts on metadata files.
 
 Quick Info:
-Displays contract information in Quick Info tool tips.
+Displays contracts in Quick Info tool tips.
 
 Signature Help:
-Displays contract information in Signature Help tool tips.</Description>
+Displays contracts in Signature Help tool tips.</Description>
     <Locale>1033</Locale>
-    <MoreInfoUrl>http://research.microsoft.com/contracts</MoreInfoUrl>
+    <MoreInfoUrl>http://visualstudiogallery.msdn.microsoft.com/1ec7db13-3363-46c9-851f-1ce455f66970</MoreInfoUrl>
     <License>AdornmentsEULA.rtf</License>
     <Icon>QuickInfo_Large.png</Icon>
     <PreviewImage>QuickInfo_Large.png</PreviewImage>
@@ -28,11 +28,11 @@ Displays contract information in Signature Help tool tips.</Description>
         <Edition>VST_All</Edition>
       </VisualStudio>
     </SupportedProducts>
-    <SupportedFrameworkRuntimeEdition MinVersion="4.0" MaxVersion="4.0" />
+    <SupportedFrameworkRuntimeEdition MinVersion="4.0" MaxVersion="5.0" />
   </Identifier>
   <References />
   <Content>
-    <VsPackage>CodeContractsHelper_Package.pkgdef</VsPackage>
-    <MefComponent>CodeContractsHelper.dll</MefComponent>
+    <VsPackage>|%CurrentProject%|</VsPackage>
+    <MefComponent>|%CurrentProject%|</MefComponent>
   </Content>
 </Vsix>
