+++ Microsoft.Research/VSAdornments/AdornmentManager.cs
