+++ Microsoft.Research/VSAdornments/ContractAdornment.xaml
