+++ Microsoft.Research/VSAdornments/ContractAdornment.xaml.cs
