+++ Microsoft.Research/VSAdornments/IAdornment.cs
