+++ Microsoft.Research/VSAdornments/IStaticAdornment.cs
