+++ Microsoft.Research/VSAdornments/InheritanceContractAdornment.cs
