+++ Microsoft.Research/VSAdornments/Logger.cs
