+++ Microsoft.Research/VSAdornments/MetadataContractAdornment.cs
