+++ Microsoft.Research/VSAdornments/Properties/AssemblyInfo.cs
