+++ Microsoft.Research/VSAdornments/PurityAdornment.cs
