+++ Microsoft.Research/VSAdornments/VSTextPropertiesProvider.cs
